Fix formatting

+++ src/SwapFacility.sol
@@ -49,8 +49,7 @@ contract SwapFacility is ISwapFacility, Owned {
 
     uint256 internal constant ORACLE_STALE_THRESHOLD = 24 hours;
     uint256 internal constant BPS = 1e4;
-    uint internal constant MAX_SPREAD = 0.1e4; // 10%
-
+    uint256 internal constant MAX_SPREAD = 0.1e4; // 10%
 
     uint256 internal constant ORACLE_STALE_THRESHOLD = 1 hours;
